+++ src/pages/TagsPage.vue
@@ -1,9 +1,9 @@
 <template>
-  <LoadError v-if="isError" :detail="errorDetail">
+  <ErrorScreen v-if="isError" :detail="errorDetail">
     <PrimaryButton button-text="Retry" @click="onRetry">
       <ArrowPathIcon class="-ml-1 mr-2 h-5 w-5" aria-hidden="true" />
     </PrimaryButton>
-  </LoadError>
+  </ErrorScreen>
   <div v-if="data">
     <div class="border-b border-gray-200 bg-white p-4">
       <div class="-ml-4 -mt-2 flex flex-nowrap items-center justify-between">
@@ -42,7 +42,7 @@
 <script setup>
 import { useTags } from '../composables/useTags'
 import { computed, ref } from 'vue'
-import LoadError from '../components/LoadError.vue'
+import ErrorScreen from '../components/ErrorScreen.vue'
 import PrimaryButton from '../components/PrimaryButton.vue'
 import { ArrowPathIcon } from '@heroicons/vue/20/solid'
