+++ src/pg/Playground.vue
@@ -1,11 +1,11 @@
 <script setup>
-import LoadError from '../components/LoadError.vue'
+import ErrorScreen from '../components/ErrorScreen.vue'
 </script>
 
 <template>
   <div
     class="mx-auto my-10 max-w-xl space-y-10 border px-4 pt-16 sm:px-6 lg:px-8"
   >
-    <LoadError />
+    <ErrorScreen />
   </div>
 </template>
